feat: #2332 added tenant base guard
